refactor: move pool-incentive components
