--- packages/web/src/components/incentivize/select-distribution-period-input/SelectDistributionPeriodInput.styles.ts
@@ -1,77 +1,0 @@
-import { fonts } from "./../../../constants/font.constant";
-import styled from "@emotion/styled";
-import mixins from "@styles/mixins";
-
-export const SelectDistributionPeriodInputWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: auto;
-  & .description {
-    ${mixins.flexbox("row", "center", "flex-start")}
-    ${fonts.p4}
-    color: ${({ theme }) => theme.color.text04};
-    margin-bottom: 4px;
-    gap: 4px;
-  }
-
-  & .period-wrapper {
-    position: relative;
-    display: flex;
-    width: 100%;
-    height: 48px;
-    padding: 0px 16px;
-    gap: 8px;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    border-radius: 8px;
-    border: 1px solid ${({ theme }) => theme.color.border02};
-    background: ${({ theme }) => theme.color.background20};
-
-    & .icon-arrow {
-      width: 16px;
-      height: 16px;
-    }
-
-    & .period {
-      height: 16px;
-      ${fonts.body9}
-      color: ${({ theme }) => theme.color.text02};
-    }
-  }
-`;
-
-export const PoolIncentivizeSelectPeriodBox = styled.div`
-  position: absolute;
-  top: 53px;
-  left: 0;
-  width: 100%;
-  background-color: ${({ theme }) => theme.color.background06};
-  border: 1px solid ${({ theme }) => theme.color.border02};
-  border-radius: 8px;
-  box-shadow: 10px 14px 60px 0px rgba(0, 0, 0, 0.4);
-`;
-
-export const PoolIncentivizeSelectPeriodBoxItem = styled.div`
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  padding: 8px 24px;
-  color: ${({ theme }) => theme.color.text02};
-  ${fonts.body12}
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  transition: 0.2s;
-
-  &:hover {
-    background-color: ${({ theme }) => theme.color.hover02};
-  }
-`;
-
-export const DistributionPeriodTooltipContentWrapper = styled.div`
-  ${fonts.body12}
-  width: 350px;
-  color: ${({ theme }) => theme.color.text11};
-`;
