+++ packages/web/src/views/pool/pool-incentivize/components/incentivize-pool-modal/IncentivizePoolModal.styles.ts
