+++ packages/web/src/views/pool/pool-incentivize/components/pool-incentivize/disclaimer/Disclaimer.tsx
