+++ packages/web/src/views/pool/pool-incentivize/components/pool-incentivize/pool-incentivize-details/incentivize-block-tooltip-content/IncentivizeBlockTooltipContent.styles.ts
