+++ packages/web/src/views/pool/pool-incentivize/components/pool-incentivize/pool-incentivize-select-pool/pool-incentivize-select-pool-item/PoolIncentivizeSelectPoolItem.styles.ts
