+++ packages/web/src/views/pool/pool-incentivize/components/pool-incentivize/select-distribution-period/SelectDistributionPeriod.styles.ts
