+++ packages/web/src/views/pool/pool-incentivize/hooks/use-incentivize-pool-modal.tsx
@@ -1,0 +1,24 @@
+import { useAtom } from "jotai";
+import { useCallback } from "react";
+
+import { CommonState } from "@states/index";
+
+import IncentivizePoolModalContainer from "../containers/incentivize-pool-modal-container/IncentivizePoolModalContainer";
+
+export interface Props {
+  openModal: () => void;
+}
+
+export const useIncentivizePoolModal = (): Props => {
+  const [, setOpenedModal] = useAtom(CommonState.openedModal);
+  const [, setModalContent] = useAtom(CommonState.modalContent);
+
+  const openModal = useCallback(() => {
+    setOpenedModal(true);
+    setModalContent(<IncentivizePoolModalContainer />);
+  }, [setModalContent, setOpenedModal]);
+
+  return {
+    openModal,
+  };
+};
